requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.4
django-allauth==64.1.0
django-jazzmin==3.0.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.2
